+++ packages/webstudio-asset-uploader/.gitignore
@@ -1,0 +1,2 @@
+lib
+generated
