+++ packages/webstudio-asset-uploader/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  testEnvironment: "node",
+  testMatch: ["<rootDir>/src/**/*.test.ts"],
+  transform: {
+    "^.+\\.tsx?$": "esbuild-jest",
+  },
+  moduleNameMapper: {
+    "^~/(.*)$": "<rootDir>/src/$1",
+  },
+};
