+++ packages/webstudio-asset-uploader/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@webstudio-is/asset-uploader",
+  "version": "0.0.1",
+  "description": "Webstudio Asset Uploader",
+  "author": "Webstudio <github@webstudio.is>",
+  "homepage": "https://webstudio.is",
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "build": "rm -fr lib && tsc",
+    "test": "jest",
+    "dev": "tsc --watch",
+    "checks": "yarn typecheck && yarn test",
+    "prepublishOnly": "yarn checks && yarn build"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.17.12",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^17.0.21",
+    "jest": "^27.5.1",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^4.6.2",
+    "@aws-sdk/client-s3": "^3.131.0",
+    "@aws-sdk/lib-storage": "^3.131.0",
+    "zod": "^3.17.3"
+  },
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "files": [
+    "README.md",
+    "!*.test.*"
+  ],
+  "license": "MIT",
+  "private": false
+}
