+++ packages/webstudio-asset-uploader/src/targets/s3/uploader.ts
@@ -1,0 +1,23 @@
+export const uploadToS3 = async ({
+  formData,
+  projectId,
+  db,
+}: {
+  formData: FormData;
+  projectId: string;
+  db: any;
+}) => {
+  try {
+    const imagesInfo = formData.getAll("image") as Array<string>;
+    const allInfo = imagesInfo.map(async (image) => {
+      const uploadedImage = JSON.parse(image);
+      const newAsset = await db.assets.createFromS3(projectId, uploadedImage);
+      return newAsset;
+    });
+    await Promise.all(allInfo);
+  } catch (error) {
+    if (error instanceof Error) {
+      throw new Error(error.message);
+    }
+  }
+};
